DATAMONGO-2384 - Introduce namespace elements for .client.MongoClient and ClientSettings.
This change introduces configuration options for the com.mongodb.client.MongoClient that will be used in the upcoming version of the MongoDB Java driver.
To fully support the client and its builders the XML namespace has been enhanced with mongo.mongo-client deprecating the mongo-client. Once all driver bits are available for the new version it is likely that the namespace has to entirely switch to mongo.mongo-client which will then be renamed to just mongo-client.